@@ -5,6 +5,8 @@ namespace QuanLyLopHoc.RequestModel
     public class LopHocRequestAdd
     {
         public string TenLopHoc { get; set; }
-        public List<int> SinhVienIds { get; set; } = new List<int>();
+        public List<SinhVienRequestAdd> SinhViens { get; set; } = new List<SinhVienRequestAdd>();
+
+        public List<BaiTapRequestAdd> BaiTaps { get; set; } = new List<BaiTapRequestAdd>();
     }
 }
